+++ mem.h
@@ -4,4 +4,4 @@ void *xcalloc(size_t n, size_t size);
 void *xrealloc(void *p, size_t n);
 
 // monotonic allocation
-void *mmalloc(int n);
+void *mmalloc(size_t n);
