+++ sym.cc
@@ -1,17 +1,17 @@
 #include "main.h"
 
 namespace {
-int cap = 0x100;
-int count;
+size_t cap = 0x100;
+size_t count;
 sym **entries = (sym **)xcalloc(cap, sizeof(sym *));
 
-bool eq(const sym *S, const char *s, int n) {
+bool eq(const sym *S, const char *s, size_t n) {
   if (S->n != n)
     return false;
   return !memcmp(S->s, s, n);
 }
 
-int slot(sym **entries, int cap, const char *s, int n) {
+size_t slot(sym **entries, size_t cap, const char *s, size_t n) {
   auto mask = cap - 1;
   auto i = fnv(s, n) & mask;
   while (entries[i] && !eq(entries[i], s, n))
@@ -22,7 +22,7 @@ int slot(sym **entries, int cap, const char *s, int n) {
 void expand() {
   auto cap1 = cap * 2;
   auto entries1 = (sym **)xcalloc(cap1, sizeof(sym *));
-  for (int i = 0; i != cap; ++i) {
+  for (size_t i = 0; i != cap; ++i) {
     auto S = entries[i];
     if (S)
       entries1[slot(entries1, cap1, S->s, S->n)] = S;
@@ -32,7 +32,7 @@ void expand() {
   entries = entries1;
 }
 
-sym *store(const char *s, int n) {
+sym *store(const char *s, size_t n) {
   auto r = (sym *)mmalloc(offsetof(sym, s) + n);
   r->n = n;
   memcpy(r->s, s, n);
@@ -41,7 +41,7 @@ sym *store(const char *s, int n) {
 
 struct init {
   init() {
-    for (int i = 0; i != nkeywords; ++i) {
+    for (size_t i = 0; i != nkeywords; ++i) {
       auto S = keywords + i;
       assert(S->n < sizeof S->s);
       ++count;
@@ -54,7 +54,7 @@ struct init {
 } init1;
 } // namespace
 
-sym *intern(const char *s, int n) {
+sym *intern(const char *s, size_t n) {
   if (n > 0xffff)
     err("Symbol too long");
   auto i = slot(entries, cap, s, n);
